Revert bug that caused rate limiter to not start on zero (on purpose)
The bug was making it so the rate limiter reinitialized to the value wired to it, but in pure pursuit we wanted it to be reinitialized to 0. The fix is to add a starting value control and that is what it will be reinitialized to. The lesson here is to not mess with very basic, important subVIs